// apps/recruiter-dashboard/__generated__/OrganizationInvitesListInternalFragment.graphql.ts
export type OrganizationInvitesListInternalFragment$data = {
  readonly id: string;
  readonly invites: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"InviteFragment">;
      };
    }>;
    readonly pageInfo: {
      readonly hasNextPage: boolean;
    };
  };
  readonly " $fragmentType": "OrganizationInvitesListInternalFragment";
};
export type OrganizationInvitesListInternalFragment$key = {
  readonly " $data"?: OrganizationInvitesListInternalFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"OrganizationInvitesListInternalFragment">;
};